Делит js код на модули
